fix the edit form
